extract event types
