+++ src/lib/ui/ui.ts
@@ -1,26 +1,21 @@
 import type {Readable, Writable} from 'svelte/store';
 import {writable, derived, get} from 'svelte/store';
 import {setContext, getContext} from 'svelte';
-import type {Static} from '@sinclair/typebox';
 
 import type {Community} from '$lib/vocab/community/community';
 import type {Space} from '$lib/vocab/space/space';
 import type {Persona} from '$lib/vocab/persona/persona';
-import type {ClientAccountSession, ClientSession} from '$lib/session/clientSession';
+import type {ClientSession} from '$lib/session/clientSession';
 import type {AccountModel} from '$lib/vocab/account/account';
 import type {File} from '$lib/vocab/file/file';
 import type {Membership} from '$lib/vocab/membership/membership';
-import type {ApiResult} from '$lib/server/api';
-import type {createCommunityService} from '$lib/vocab/community/communityServices';
-import type {createPersonaService} from '$lib/vocab/persona/personaServices';
-import type {createMembershipService} from '$lib/vocab/community/communityServices';
-import type {createSpaceService} from '$lib/vocab/space/spaceServices';
-import type {createFileService, readFilesService} from '$lib/vocab/file/fileServices';
-import type {Dispatch, DispatchContext} from '$lib/ui/api';
-import type {LoginRequest} from '$lib/session/loginMiddleware.js';
+import type {DispatchContext} from '$lib/ui/api';
+import type {UiHandlers} from '$lib/ui/events';
 
 const UNKNOWN_API_ERROR =
 	'Something went wrong. Maybe the server or your Internet connection is down. Please try again.';
+
+export type MainNavView = 'explorer' | 'account'; // TODO where should this live?
 
 const KEY = Symbol();
 
@@ -30,65 +25,6 @@ export const setUi = (store: Ui): Ui => {
 	setContext(KEY, store);
 	return store;
 };
-
-// TODO generate this interface from data or define programmatically along with `Dispatch`
-export interface UiHandlers {
-	// TODO `log_in` and `log_out` have no service or schema; they're custom endpoints right now,
-	// but they should be converted to services to match the rest.
-	// For now they serve as good examples of a usecase needing an escape hatch,
-	// because they need to use the api client directly. (because `invoke` is only for services)
-	log_in: (
-		ctx: DispatchContext<LoginRequest, ApiResult<{session: ClientAccountSession}>>,
-	) => Promise<ApiResult<{session: ClientAccountSession}>>;
-	log_out: (ctx: DispatchContext<null, ApiResult<void>>) => Promise<ApiResult<void>>;
-	create_community: (
-		ctx: DispatchContext<
-			Static<typeof createCommunityService.paramsSchema>,
-			ApiResult<Static<typeof createCommunityService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof createCommunityService.responseSchema>>>;
-	create_persona: (
-		ctx: DispatchContext<
-			Static<typeof createPersonaService.paramsSchema>,
-			ApiResult<Static<typeof createPersonaService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof createPersonaService.responseSchema>>>;
-	create_membership: (
-		ctx: DispatchContext<
-			Static<typeof createMembershipService.paramsSchema>,
-			ApiResult<Static<typeof createMembershipService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof createMembershipService.responseSchema>>>;
-	create_space: (
-		ctx: DispatchContext<
-			Static<typeof createSpaceService.paramsSchema>,
-			ApiResult<Static<typeof createSpaceService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof createSpaceService.responseSchema>>>;
-	create_file: (
-		ctx: DispatchContext<
-			Static<typeof createFileService.paramsSchema>,
-			ApiResult<Static<typeof createFileService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof createFileService.responseSchema>>>;
-	read_files: (
-		ctx: DispatchContext<
-			Static<typeof readFilesService.paramsSchema>,
-			ApiResult<Static<typeof readFilesService.responseSchema>>
-		>,
-	) => Promise<ApiResult<Static<typeof readFilesService.responseSchema>>>;
-	query_files: (
-		ctx: DispatchContext<Static<typeof readFilesService.paramsSchema>>,
-		dispatch: Dispatch,
-	) => Readable<Readable<File>[]>;
-	toggle_main_nav: () => void;
-	toggle_secondary_nav: () => void;
-	set_main_nav_view: (ctx: DispatchContext<MainNavView>) => void;
-	set_mobile: (ctx: DispatchContext<boolean>) => void;
-	select_persona: (ctx: DispatchContext<{persona_id: number}>) => void;
-	select_community: (ctx: DispatchContext<{community_id: number | null}>) => void;
-	select_space: (ctx: DispatchContext<{community_id: number; space_id: number | null}>) => void;
-}
 
 export interface Ui extends UiHandlers {
 	dispatch: (ctx: DispatchContext) => any; // TODO return value type?
@@ -555,8 +491,6 @@ export const toUi = (session: Writable<ClientSession>, initialMobile: boolean): 
 	};
 	return ui;
 };
-
-export type MainNavView = 'explorer' | 'account';
 
 // TODO this is a hack until we have `community_ids` normalized and off the `Persona`,
 // the issue is that the "session personas" are different than the rest of the personas
